+++ lib/app/data/storage/data/data.dart
@@ -31,7 +31,7 @@ import 'package:money/app/data/models/money_objects/transfers/transfer.dart';
 import 'package:money/app/data/models/settings.dart';
 import 'package:money/app/data/storage/database/database.dart';
 import 'package:money/app/core/widgets/snack_bar.dart';
-import 'package:money/app/routes/home_data_controller.dart';
+import 'package:money/app/modules/home/home_data_controller.dart';
 
 // Exports
 export 'package:money/app/core/helpers/json_helper.dart';