+++ lib/main.dart
@@ -2,10 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:money/app/core/theme/theme_controler.dart';
 import 'package:money/app/data/models/settings.dart';
-import 'package:money/app/routes/home_data_controller.dart';
+import 'package:money/app/modules/general/general_routes.dart';
+import 'package:money/app/modules/home/home_data_controller.dart';
+import 'package:money/app/modules/home/home_routes.dart';
+import 'package:money/app/modules/policies/policy_routes.dart';
+import 'package:money/app/modules/splash_screen.dart';
+import 'package:money/app/modules/welcome/welcome_page.dart';
+import 'package:money/app/modules/welcome/welcome_routes.dart';
 
 import 'app/core/bindings/application_bindings.dart';
-import 'app/routes/app_pages.dart';
 
 void main() {
   runApp(MyApp());
@@ -38,7 +43,21 @@ class MyApp extends StatelessWidget {
           title: 'MyMoney by VTeam',
           initialBinding: ApplicationBindings(),
           initialRoute: '/',
-          getPages: AppPages.routes,
+          getPages: [
+            GetPage(
+                name: '/',
+                page: () {
+                  PreferenceController preferenceController = Get.find();
+                  if (preferenceController.isReady.value) {
+                    return const WelcomePage();
+                  }
+                  return const SplashScreen();
+                }),
+            ...HomeRoutes.routes,
+            ...WelcomeRoutes.routes,
+            ...GeneralRoutes.routes,
+            ...PolicyRoutes.routes,
+          ],
         );
       },
     );